requirements.txt
streamlit
chromadb
httpx
python-dotenv
